Moved email preview spec to acceptance tests
- Tests here are baseline usecases which should be always in check
- More edge case scenarios might go into regression suite in the future
