+++ core/test/regression/api/canary/admin/utils.js
@@ -71,7 +71,7 @@ const expectedProperties = {
     webhook: _(schema.webhooks)
         .keys()
     ,
-    email_preview: ['html', 'subject']
+    email_preview: ['html', 'subject', 'plaintext']
 };
 
 _.each(expectedProperties, (value, key) => {
